pyproject.toml
[tool.poetry]
name = "abcfood-mm-core"
version = "0.1.0"
description = "Core business logic service for Mattermost command center - approvals, ChatOps, audit logging"
authors = ["ABCFood Team <dev@abcfood.app>"]
readme = "README.md"
license = "Proprietary"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
python-dotenv = "^1.0.0"
structlog = "^24.4.0"
httpx = "^0.28.0"
psycopg2-binary = "^2.9.0"
clickhouse-connect = "^0.8.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
cachetools = "^5.3.0"
python-multipart = "^0.0.18"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
black = "^24.10.0"

[tool.poetry.scripts]
mm-core = "app.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=term-missing"
